backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
google-genai==1.73.1
google-cloud-logging==3.12.1
google-cloud-secret-manager==2.22.1
google-cloud-firestore==2.19.0
httpx==0.28.1
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
pypdf==4.2.0
